requirements.txt
python-dotenv==1.0.0
Flask==2.3.3
playwright>=1.54.0
openai>=1.30.0